Postponing this issue for now. Fixed supabase duplicate rows.
