+++ action/loadData.ts
@@ -1,9 +1,14 @@
-import { FetchRiskData, FetchSchoolData } from "@/app/api/data/downloadSchoolData";
-import useMidasStore from "@/hooks/useSchoolData";
+import { FetchSchoolData } from "@/app/api/data/downloadSchoolData";
+import { FetchRiskData } from "@/app/api/data/risk-data/route";
+import useMidasStore, { SchoolData } from "@/hooks/useSchoolData";
 
 
 export async function loadData(schoolId: number) {
   try {
+    //const resSchoolData = await fetch(`/api/data/school-data/school_id=${schoolId}`);
+    //const jsonSchoolData = await resSchoolData.json();
+    //const schoolData = jsonSchoolData.data;
+    //
     const schoolData = await FetchSchoolData(schoolId);
     const riskData = await FetchRiskData(schoolId);
 
@@ -11,8 +16,8 @@ export async function loadData(schoolId: number) {
       throw new Error("Failed to load data");
     }
 
-    const combinedData = schoolData.map(student => {
-      const studentRisks = riskData.filter(risk => risk.studentid === student.studentid);
+    const combinedData = schoolData.map((student: SchoolData) => {
+      const studentRisks = riskData.filter((risk: any) => risk.studentid === student.studentid);
       console.log("STUDENT RISKS");
       console.log(studentRisks);
       const riskObj = {
@@ -21,7 +26,7 @@ export async function loadData(schoolId: number) {
         midas: {}
       };
 
-      studentRisks.forEach(risk => {
+      studentRisks.forEach((risk: any) => {
         console.log(`RISK N: ${risk.factor}`)
         if (risk.factor.toLowerCase() === 'student') {
           riskObj.student = {
